updates packages, quiets logging output
